Move goto outside of ReachySDK.
